requirements.txt
fastapi
uvicorn
google-genai
pandas
openpyxl
sqlalchemy
psycopg2-binary